fix(ui): working, consistent copy buttons in detail tabs; filter top spacing
- add shared clipboard util with a non-secure-context fallback; use it in the
  container row, Container tab, and Image tab (replaces the navigator-only copies)
- Image-tab copy buttons move beside their value to match the Container tab, in
  the muted text colour
- add breathing room above the container filter bar
